Fix #705

+++ code/modules/antagonist/contracts/kill.dm
@@ -52,6 +52,7 @@
 /datum/contract/kill/proc/get_target()
 	var/datum/mind/list/candidates = list()
 	for(var/datum/mind/M in (ticker.minds - get_taken_targets()))
+		if((M.assigned_role == "Gladiator"))	continue // no thunderdome gladiators
 		if((M.assigned_role in command_positions))	continue // head contract's stuff
 		if((M in faction.members) || (M.antagonist && (M.antagonist.faction.name in faction.alliances)))	continue // no killing coworkers or allies. heads are excluded from normal kill contracts
 		if(ishuman(M.current) && M.current.stat != 2)
